requirements.txt
# ===============================
# Web interface (HTML / API)
# ===============================
flask>=2.3.0
flask-cors>=4.0.0
jinja2>=3.1.0

# ===============================
# Data handling
# ===============================
pandas>=2.0.0
numpy>=1.24.0

# ===============================
# PDF & text processing
# ===============================
pymupdf>=1.23.0        # PDF -> text (fitz)
pdfplumber>=0.10.0
python-docx>=1.1.0
tqdm>=4.66.0

# ===============================
# NLP / Embeddings / RAG
# ===============================
sentence-transformers>=2.2.2
transformers>=4.38.0
torch>=2.0.0
faiss-cpu>=1.7.4       # use faiss-gpu if CUDA available

# ===============================
# LLM clients (choose what you use)
# ===============================
openai>=1.10.0         # optional (API)
huggingface-hub>=0.20.0

# ===============================
# Database (SQLite)
# ===============================
sqlalchemy>=2.0.0

# ===============================
# Utilities
# ===============================
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
